+++ src/lib/components/custom/sidebar/add-panel-and-view-trigger.svelte
@@ -19,7 +19,6 @@
 		parent_id,
 		highest_unit,
 		panel_name,
-		is_parent_root_node = false,
 		open_dialog_state = $bindable(false),
 		latest_circuit_node
 	}: {
@@ -30,7 +29,6 @@
 		generic_phase_panel_form: SuperValidated<GenericPhasePanelSchema>;
 		parent_id: string;
 		open_dialog_state?: boolean;
-		is_parent_root_node: boolean;
 		latest_circuit_node?: Node;
 	} = $props();