+++ src/lib/hooks/settings-state.svelte.ts
@@ -3,20 +3,24 @@ import type { Settings } from '@/types/settings';
 import setGlobalColorTheme from '@/theme-colors/theme-colors';
 import { setState, getState } from '@/state/index.svelte';
 import { THEME_COLOR_STATE_CTX } from '@/state/constants';
+import { setMode } from 'mode-watcher';
 
 export type Font = 'isocpeur' | 'verdana' | 'default';
 type ThemeColor = 'autocad' | 'excel';
-type ThemeMode = 'dark' | 'light' | 'system';
+export type ThemeMode = 'dark' | 'light' | 'system';
 
 export class SettingsState {
-	persisted_state: PersistedState<Settings>;
+	private persisted_state: PersistedState<Settings>;
 
 	themeColor = $state<ThemeColor>('excel');
+	themeMode = $state<ThemeMode>('light');
 	font = $state<Font>('default');
 	show_loads_on_unit_hierarchy = $state(false);
 	is_adjustment_factor_dynamic = $state(false);
 	is_panel_multi_copy = $state(false);
 	is_load_multi_copy = $state(false);
+	auto_save_enabled = $state(false);
+	backup_project_file_if_exists = $state(false);
 
 	constructor(mode: ThemeMode) {
 		const _persisted_state = new PersistedState<Settings>('settings', {
@@ -26,15 +30,26 @@ export class SettingsState {
 			is_adjustment_factor_dynamic: false,
 			is_panel_multi_copy: false,
 			is_load_multi_copy: false,
+			theme_mode: mode,
+			auto_save_enabled: false,
+			backup_project_file_if_exists: false
 		});
 
 		// must be set before calling setters
 		this.persisted_state = _persisted_state;
-
-		this.setThemeColor(_persisted_state?.current?.color || 'excel', mode);
+		this.setAutoSave(_persisted_state?.current?.auto_save_enabled || false);
+		this.setThemeColor(
+			_persisted_state?.current?.color || 'excel',
+			_persisted_state.current.theme_mode || 'light'
+		);
+		this.setThemeMode(_persisted_state.current.theme_mode || 'light');
 		this.setGlobalFont(_persisted_state?.current?.font ?? 'default');
-		this.setShowLoadsOnUnitHeirarchy(_persisted_state?.current?.show_loads_on_unit_hierarchy || false);
-		this.setIsAdjustmentFactorDynamic(_persisted_state?.current?.is_adjustment_factor_dynamic || false);
+		this.setShowLoadsOnUnitHeirarchy(
+			_persisted_state?.current?.show_loads_on_unit_hierarchy || false
+		);
+		this.setIsAdjustmentFactorDynamic(
+			_persisted_state?.current?.is_adjustment_factor_dynamic || false
+		);
 		this.setIsPanelMultiCopy(_persisted_state?.current?.is_panel_multi_copy || false);
 		this.setIsLoadMultiCopy(_persisted_state?.current?.is_load_multi_copy || false);
 	}
@@ -42,6 +57,17 @@ export class SettingsState {
 	private setGlobalFont(font: Font) {
 		if (typeof document !== 'undefined') {
 			document.body.className = font;
+		}
+	}
+
+	private setGlobalThemeMode(mode: ThemeMode) {
+		const systemPrefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+		if (mode === 'system') {
+			setGlobalColorTheme(systemPrefersDark ? 'dark' : 'light', this.themeColor);
+			setMode(systemPrefersDark ? 'dark' : 'light');
+		} else {
+			setGlobalColorTheme(mode, this.themeColor);
+			setMode(mode);
 		}
 	}
 
@@ -54,12 +80,37 @@ export class SettingsState {
 		setGlobalColorTheme(mode, color);
 	}
 
+	setAutoSave(value: boolean) {
+		this.auto_save_enabled = value;
+		this.persisted_state.current = {
+			...this.persisted_state.current,
+			auto_save_enabled: value
+		};
+	}
+
+	setBackupProjectFileIfExists(value: boolean) {
+		this.backup_project_file_if_exists = value;
+		this.persisted_state.current = {
+			...this.persisted_state.current,
+			backup_project_file_if_exists: value
+		};
+	}
+
 	setFont(font: Font) {
 		this.font = font;
 		this.setGlobalFont(font);
 		this.persisted_state.current = {
 			...this.persisted_state.current,
 			font
+		};
+	}
+
+	setThemeMode(mode: ThemeMode) {
+		this.themeMode = mode;
+		this.setGlobalThemeMode(mode);
+		this.persisted_state.current = {
+			...this.persisted_state.current,
+			theme_mode: mode
 		};
 	}